+++ choreographer/browsers/__init__.py
@@ -1,4 +1,4 @@
-"""browsers contains implementations of browsers that choreographer can open."""
+"""Contains implementations of browsers that choreographer can open."""
 
 from ._errors import BrowserClosedError, BrowserFailedError
 from .chromium import Chromium
