+++ choreographer/channels/__init__.py
@@ -1,4 +1,9 @@
-"""Channels are classes that choreo and the browser use to communicate."""
+"""
+Channels are classes that choreo and the browser use to communicate.
+
+This is a low-level part of the API.
+
+"""
 
 from ._errors import BlockWarning, ChannelClosedError, JSONError
 from .pipe import Pipe
