--- choreographer/cli/README.txt
@@ -1,1 +1,0 @@
-CLI tools provide various functions that are turned into scripts.
